requirements.txt
pandas
scikit-learn
streamlit
joblib
pytesseract
Pillow
